feat(82): dmn decsionDefinition storage + refactor runtime pkg
refactor bpmnruntime, dmnruntime packages, remove int32 restrictions from DB

+++ internal/sql/job.sql.go
@@ -259,8 +259,8 @@ WHERE
 `
 
 type FindProcessInstanceJobsInStateParams struct {
-	ProcessInstanceKey int64 `json:"process_instance_key"`
-	States             []int `json:"states"`
+	ProcessInstanceKey int64   `json:"process_instance_key"`
+	States             []int64 `json:"states"`
 }
 
 func (q *Queries) FindProcessInstanceJobsInState(ctx context.Context, arg FindProcessInstanceJobsInStateParams) ([]Job, error) {
@@ -322,7 +322,7 @@ type SaveJobParams struct {
 	ElementInstanceKey int64  `json:"element_instance_key"`
 	ProcessInstanceKey int64  `json:"process_instance_key"`
 	Type               string `json:"type"`
-	State              int    `json:"state"`
+	State              int64  `json:"state"`
 	CreatedAt          int64  `json:"created_at"`
 	Variables          string `json:"variables"`
 	ExecutionToken     int64  `json:"execution_token"`
